src/main/java/uk/nhs/prm/deductions/pdsadaptor/service/PdsService.java
package uk.nhs.prm.deductions.pdsadaptor.service;

import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;
import uk.nhs.prm.deductions.pdsadaptor.client.PdsFhirClient;
import uk.nhs.prm.deductions.pdsadaptor.model.pdsresponse.PdsResponse;

@Service
@RequiredArgsConstructor
public class PdsService {

    private final PdsFhirClient pdsFhirClient;

    public PdsResponse getPatientGpStatus(String nhsNumber) {
        return pdsFhirClient.requestPdsRecordByNhsNumber(nhsNumber);
    }
}
